--- src/HasViews.php
@@ -1,19 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-/*
- * This file is part of the Eloquent Viewable package.
- *
- * (c) Cyril de Wit <github@cyrildewit.nl>
- *
- * For the full copyright and license information, please view the LICENSE
- * file that was distributed with this source code.
- */
-
-namespace CyrildeWit\EloquentViewable;
-
-
-interface HasViews
-{
-}
